reorganized code with both formpix and the sim
separated the app into modules for better clarity and organization
