+++ formPixSim/.env-template
@@ -1,0 +1,8 @@
+formbarUrl = http://localhost:420
+api = put-your-api-key-here
+brightness = 0
+pin = 0
+stripType = WS2812
+barPixels = 0
+boards = 0
+port = 421
