--- settings-template.json
@@ -1,10 +1,0 @@
-{
-	"formbarUrl": "",
-	"api": "",
-	"brightness": 0,
-	"pin": 0,
-	"stripType": "",
-	"barPixels": 0,
-	"boards": 0,
-	"port": 421
-}
